+++ clang-tools-extra/clang-tidy/bsltest/BslTestTidyModule.cpp
@@ -1,4 +1,4 @@
-//===------- BoostTidyTestModule.cpp - clang-tidy -----------------------------===//
+//===------- BslTestTidyTestModule.cpp - clang-tidy -----------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -10,28 +10,34 @@
 #include "../ClangTidyModule.h"
 #include "../ClangTidyModuleRegistry.h"
 #include "BitwiseTypeCheck.h"
+#include "CastSignednessCheck.h"
+#include "ForLoopCheck.h"
 using namespace clang::ast_matchers;
 
 namespace clang {
 namespace tidy {
-namespace boost {
+namespace bsltest {
 
 class BslTestModule : public ClangTidyModule {
 public:
   void addCheckFactories(ClangTidyCheckFactories &CheckFactories) override {
     CheckFactories.registerCheck<BitwiseTypeCheck>(
         "bsltest-bitwise-type");
+    CheckFactories.registerCheck<CastSignednessCheck>(
+        "bsltest-cast-signedness");
+    CheckFactories.registerCheck<ForLoopCheck>(
+        "bsltest-for-loop");
   }
 };
 
-// Register the BoostModule using this statically initialized variable.
+// Register the BslTestModule using this statically initialized variable.
 static ClangTidyModuleRegistry::Add<BslTestModule> X("bsl-test-module",
                                                    "Add additional bsl checks.");
 
-} // namespace boost
+} // namespace bsltest
 
 // This anchor is used to force the linker to link in the generated object file
-// and thus register the BoostModule.
+// and thus register the BslTestModule.
 volatile int BslTestModuleAnchorSource = 0;
 
 } // namespace tidy